Moved Bottom Nav bar into its own place for components
